+++ src/main/java/com/study/realworld/global/config/WebSecurityConfig.java
@@ -2,8 +2,8 @@ package com.study.realworld.global.config;
 
 import static org.springframework.http.HttpMethod.POST;
 
-import com.study.realworld.jwt.JwtAuthenticationTokenFilter;
-import com.study.realworld.jwt.JwtTokenProvider;
+import com.study.realworld.security.JwtAuthenticationTokenFilter;
+import com.study.realworld.security.JwtTokenProvider;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
